requirements.txt
web3==6.11.1
pycryptodome==3.19.0
py-ecc==7.0.0
eth-account==0.8.0
eth-typing==3.5.1
eth-utils==2.3.1
python-dotenv==1.0.0
fastapi==0.109.1
uvicorn==0.27.0
pydantic==2.6.1
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
multidict==6.0.4
yarl==1.9.3 